# Move the IsLeft/IsRight decision out of the loop