While reviewing https://github.com/dotnet/runtime/pull/87510, I noticed the inline code in the Comparers seems like since it's in the hot-loop path, might be faster to move the IsLeft conditionalized code out of the loop by adding a Slicer to the comparator. The Slicer is the same logic as was in the bodies of the Equals and GetHashCode methods, and indeed also matches the delegate that was passed to the internal CreateAnalysisResults method.

The slicing is really only changed once per Count, so move the `IsLeft`-dependent logic into aggressively inlined `SliceLeft` and `SliceRight` to see if that makes things faster. Also reuses the same slicers for the calls to the `CreateAnalysisResults` so the will get the same JIT perf benefits.

Also made  a [subtle tweak](https://github.com/dotnet/runtime/compare/main...IDisposable:dotnet-runtime:faster-freeze-strings#diff-8dfef3756fbf6ca0aab7668e6abf54ae6c416c60cb1b3bd86455204942f7dbefL127-R111) because we know that `ignoreCase` is true (due to the [test](https://github.com/dotnet/runtime/compare/main...IDisposable:dotnet-runtime:faster-freeze-strings#diff-8dfef3756fbf6ca0aab7668e6abf54ae6c416c60cb1b3bd86455204942f7dbefL127-R111) above), so can set the starting state of `canSwitchIgnoreCaseToCaseSensitive` explicitly to `true`.

``` ini

BenchmarkDotNet=v0.13.2.2052-nightly, OS=Windows 11 (10.0.22631.1972)
Intel Core i7-10875H CPU 2.30GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK=8.0.100-preview.7.23322.33
  [Host]     : .NET 8.0.0 (8.0.23.32106), X64 RyuJIT AVX2
  Job-ZWELZX : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2
  Job-KUEFVA : .NET 8.0.0 (42.42.42.42424), X64 RyuJIT AVX2

PowerPlanMode=00000000-0000-0000-0000-000000000000  Arguments=/p:EnableUnsafeBinaryFormatterSerialization=true  IterationTime=250.0000 ms  
MaxIterationCount=20  MinIterationCount=15  WarmupCount=1  

```
|                               Method |        Job |                                                                                                          Toolchain | Count |            Mean |         Error |        StdDev |          Median |             Min |             Max | Ratio | MannWhitney(3ms) | RatioSD |     Gen0 |     Gen1 |     Gen2 | Allocated | Alloc Ratio |
|------------------------------------- |----------- |------------------------------------------------------------------------------------------------------------------- |------ |----------------:|--------------:|--------------:|----------------:|----------------:|----------------:|------:|----------------- |--------:|---------:|---------:|---------:|----------:|------------:|
|                   **ToFrozenDictionary** | **Job-ZWELZX** |          **\runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe** |    **10** |     **2,057.56 ns** |     **23.550 ns** |     **20.877 ns** |     **2,053.70 ns** |     **2,004.57 ns** |     **2,095.98 ns** |  **1.26** |             **Same** |    **0.03** |   **0.2027** |        **-** |        **-** |    **1720 B** |        **1.00** |
|                   ToFrozenDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |    10 |     1,644.78 ns |     32.011 ns |     35.580 ns |     1,631.83 ns |     1,603.70 ns |     1,715.13 ns |  1.00 |             Base |    0.00 |   0.2047 |        - |        - |    1720 B |        1.00 |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|    TryGetValue_True_FrozenDictionary | Job-ZWELZX |          \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |    10 |        63.10 ns |      0.752 ns |      0.628 ns |        63.24 ns |        61.43 ns |        63.91 ns |  1.00 |             Same |    0.02 |        - |        - |        - |         - |          NA |
|    TryGetValue_True_FrozenDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |    10 |        63.30 ns |      0.842 ns |      0.788 ns |        63.31 ns |        62.34 ns |        65.07 ns |  1.00 |             Base |    0.00 |        - |        - |        - |         - |          NA |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|                   **ToFrozenDictionary** | **Job-ZWELZX** |          **\runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe** |   **100** |    **11,800.30 ns** |    **139.610 ns** |    **123.760 ns** |    **11,764.33 ns** |    **11,681.12 ns** |    **12,075.98 ns** |  **1.26** |             **Same** |    **0.01** |   **1.4098** |   **0.0470** |        **-** |   **12112 B** |        **1.00** |
|                   ToFrozenDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |   100 |     9,348.64 ns |     91.765 ns |     76.628 ns |     9,367.20 ns |     9,209.04 ns |     9,481.99 ns |  1.00 |             Base |    0.00 |   1.4145 |   0.0372 |        - |   12112 B |        1.00 |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|    TryGetValue_True_FrozenDictionary | Job-ZWELZX |          \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |   100 |       647.76 ns |     12.870 ns |     12.640 ns |       645.45 ns |       630.03 ns |       677.18 ns |  1.00 |             Same |    0.02 |        - |        - |        - |         - |          NA |
|    TryGetValue_True_FrozenDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |   100 |       645.03 ns |      8.896 ns |      8.321 ns |       645.45 ns |       631.23 ns |       658.04 ns |  1.00 |             Base |    0.00 |        - |        - |        - |         - |          NA |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|                   **ToFrozenDictionary** | **Job-ZWELZX** |          **\runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe** |  **1000** |   **107,105.57 ns** |  **1,078.411 ns** |    **955.984 ns** |   **106,742.36 ns** |   **106,269.73 ns** |   **109,294.56 ns** |  **1.26** |             **Same** |    **0.01** |  **11.1301** |   **2.5685** |        **-** |   **94864 B** |        **1.00** |
|                   ToFrozenDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |  1000 |    84,805.63 ns |  1,384.405 ns |  1,294.973 ns |    84,801.32 ns |    82,494.28 ns |    87,164.12 ns |  1.00 |             Base |    0.00 |  11.2434 |   2.6455 |        - |   94864 B |        1.00 |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|    TryGetValue_True_FrozenDictionary | Job-ZWELZX |          \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |  1000 |     7,823.96 ns |    106.727 ns |     99.833 ns |     7,841.29 ns |     7,668.40 ns |     7,999.15 ns |  1.04 |             Same |    0.02 |        - |        - |        - |         - |          NA |
|    TryGetValue_True_FrozenDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |  1000 |     7,542.00 ns |    104.779 ns |     92.883 ns |     7,555.03 ns |     7,364.24 ns |     7,702.98 ns |  1.00 |             Base |    0.00 |        - |        - |        - |         - |          NA |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|                   **ToFrozenDictionary** | **Job-ZWELZX** |          **\runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe** | **10000** | **1,641,164.67 ns** | **23,876.639 ns** | **22,334.223 ns** | **1,633,353.12 ns** | **1,612,245.00 ns** | **1,676,639.38 ns** |  **1.18** |             **Same** |    **0.02** | **143.7500** | **143.7500** | **143.7500** |  **926133 B** |        **1.00** |
|                   ToFrozenDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe | 10000 | 1,393,440.90 ns | 17,878.950 ns | 16,723.981 ns | 1,394,471.88 ns | 1,360,370.31 ns | 1,421,018.75 ns |  1.00 |             Base |    0.00 | 145.8333 | 145.8333 | 145.8333 |  926134 B |        1.00 |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|    TryGetValue_True_FrozenDictionary | Job-ZWELZX |          \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe | 10000 |   129,167.96 ns |  2,134.712 ns |  1,996.811 ns |   129,264.68 ns |   124,220.63 ns |   132,173.42 ns |  1.01 |             Same |    0.02 |        - |        - |        - |         - |          NA |
|    TryGetValue_True_FrozenDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe | 10000 |   128,107.54 ns |  1,706.036 ns |  1,595.827 ns |   127,981.50 ns |   126,341.48 ns |   131,397.43 ns |  1.00 |             Base |    0.00 |        - |        - |        - |         - |          NA |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|                         **ToDictionary** | **Job-ZWELZX** |          **\runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe** |    **10** |       **138.87 ns** |      **1.287 ns** |      **1.204 ns** |       **138.58 ns** |       **136.88 ns** |       **141.00 ns** |  **0.94** |             **Same** |    **0.02** |   **0.0526** |        **-** |        **-** |     **440 B** |        **1.00** |
|                         ToDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |    10 |       147.84 ns |      2.518 ns |      2.355 ns |       147.02 ns |       145.32 ns |       153.22 ns |  1.00 |             Base |    0.00 |   0.0526 |        - |        - |     440 B |        1.00 |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|                ToImmutableDictionary | Job-ZWELZX |          \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |    10 |     1,299.27 ns |     83.702 ns |     96.391 ns |     1,250.37 ns |     1,202.89 ns |     1,479.53 ns |  1.12 |             Same |    0.08 |   0.0860 |        - |        - |     736 B |        1.00 |
|                ToImmutableDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |    10 |     1,178.81 ns |     16.472 ns |     15.407 ns |     1,171.80 ns |     1,158.97 ns |     1,204.85 ns |  1.00 |             Base |    0.00 |   0.0847 |        - |        - |     736 B |        1.00 |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|          TryGetValue_True_Dictionary | Job-ZWELZX |          \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |    10 |       128.08 ns |      1.673 ns |      1.483 ns |       128.16 ns |       125.29 ns |       130.57 ns |  1.00 |             Same |    0.01 |        - |        - |        - |         - |          NA |
|          TryGetValue_True_Dictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |    10 |       128.21 ns |      1.885 ns |      1.763 ns |       128.02 ns |       125.36 ns |       131.44 ns |  1.00 |             Base |    0.00 |        - |        - |        - |         - |          NA |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
| TryGetValue_True_ImmutableDictionary | Job-ZWELZX |          \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |    10 |       164.46 ns |      2.012 ns |      1.882 ns |       164.70 ns |       162.56 ns |       168.03 ns |  1.00 |             Same |    0.01 |        - |        - |        - |         - |          NA |
| TryGetValue_True_ImmutableDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |    10 |       164.62 ns |      2.097 ns |      1.962 ns |       163.06 ns |       162.89 ns |       167.53 ns |  1.00 |             Base |    0.00 |        - |        - |        - |         - |          NA |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|                         **ToDictionary** | **Job-ZWELZX** |          **\runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe** |   **100** |     **1,097.04 ns** |     **21.699 ns** |     **20.298 ns** |     **1,095.36 ns** |     **1,072.24 ns** |     **1,134.88 ns** |  **1.13** |             **Same** |    **0.03** |   **0.3714** |   **0.0043** |        **-** |    **3128 B** |        **1.00** |
|                         ToDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |   100 |       972.18 ns |     16.163 ns |     15.119 ns |       968.03 ns |       958.70 ns |     1,002.61 ns |  1.00 |             Base |    0.00 |   0.3706 |   0.0039 |        - |    3128 B |        1.00 |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|                ToImmutableDictionary | Job-ZWELZX |          \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |   100 |    17,821.67 ns |    327.754 ns |    306.582 ns |    17,853.37 ns |    17,381.97 ns |    18,333.97 ns |  0.95 |             Same |    0.02 |   0.7301 |        - |        - |    6496 B |        1.00 |
|                ToImmutableDictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |   100 |    18,780.61 ns |    259.705 ns |    230.222 ns |    18,796.94 ns |    18,384.63 ns |    19,208.59 ns |  1.00 |             Base |    0.00 |   0.7449 |        - |        - |    6496 B |        1.00 |
|                                      |            |                                                                                                                    |       |                 |               |               |                 |                 |                 |       |                  |         |          |          |          |           |             |
|          TryGetValue_True_Dictionary | Job-ZWELZX |          \runtime\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |   100 |     1,442.70 ns |     19.214 ns |     17.973 ns |     1,442.10 ns |     1,410.93 ns |     1,475.64 ns |  0.95 |             Same |    0.01 |        - |        - |        - |         - |          NA |
|          TryGetValue_True_Dictionary | Job-KUEFVA | \runtime_baseline\artifacts\bin\testhost\net8.0-windows-Release-x64\shared\Microsoft.NETCore.App\8.0.0\CoreRun.exe |   100 |     1,524.14 ns |     21.632 ns |     18.064 ns |     1,521.86 ns |     1,490.48 ns |     1,565.28 ns |  1.00 |             Base |    0.00 |        - |        - |        - |         - |          NA |
…